backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
python-decouple>=3.8,<4.0
reportlab>=4.0,<5.0
dj-database-url>=2.2,<3.0
gunicorn>=22.0,<24.0
psycopg[binary]>=3.1,<4.0
whitenoise>=6.6,<7.0
